feat: Back button fix
Fixed base fragment back button interaction. Pressing refresh button was stacking instances of base fragment on top of each other such that one would have to press the back button an additional time for each time the refresh button was touched before exiting the app. MainActivity.OnBackPressed was implemented in such a fashion that now regardless of how many times the refresh button is pressed, if the baseFragment is visible the app will exit after clicking the back button once.