fix: Button during frag navigation
Button activated states now reflect the current state of the objects they represent from MainActivity. So activating a button in basefrag, then moving to foodselectfragment, then refreshing to base frag should show that the button is still activated. This eliminates the mismatch between button states and PriceRange and QuickEUsage activation states. Also added checks for making sure that there is a QuickEUsage selection and a PriceRange selection before moving to FoodSelectFragment. If there is no selection, a toast appears. However, there is proper exception recording for the case that we move forward with no option selected