feat: BrowseFragment
Implemented the core logic behind BrowseFragment. This means creating the base layout file for BrowseFragment which revolves mostly around the SwipeFlingAdapterView. A custom adapter was made for the swipeflingadapterview that loads in data from the BusinessList. A broadcast receiver and broadcast was added so that the adapter is notified of the items being added to the business list, and this processe's logic involves hiding and showing a progress bar. The actual item that swipeflingadapter is holding hasn't been beautified to show all the business information, but the key logic of presenting business items to the adapter and adapter view are all there.