bug: BusinessList population, Category selection
Fixed two major bugs. The first one involved how BusinessList was populated. There was some trouble transitioning between MainActivity and BrowseActivity and loading the data, so now the sync intent service stores all of the new businesses into a temporary ArrayList and when they are all loaded they are copied to the real BusinessList and after the copy the adapter is notified of the change via broadcast. For the most part, this allows for better transferal of data and allows us to randomize each offset of 50 as a group. Secondly, I believe it was FoodSelectFragment that was calling MainActivity.MealSelection.clear() onStop() and because of this our categories would get wiped out. The first call to YelpSync would have the categories but all subsequent calls would not. Now, the only time the clear() method is called is when basefragment calls onStart() (so basically whenever the app launches or we refresh/restart our search queries), so now all calls to YelpSync that should be the same will have the same categories.