bug: Backwards compatibility
Application was crashing on APIs older than version 21 because of ripple effectsin buttons and image buttons. To avoid this, seperate layout files were made. For v21 and up the ripple effect is still referenced, but in versions lower the ripple effect is ignored and instead the background file that just controls color is the only thing referenced. This change was made for all buttons up to BrowseFragment. Also, a try catch was added to the yes and no button onClickListeners in BrowseFragment so we avoid crashes that occured on their click due to NullPointerExcpetions occuring when TopCardListener was called without any items in the FlingContainer